Turn on the apodized FFT interpolation
The original Goddard code effectively had this turned off,
so in order to replicate those results this was turned off here
too even though was part of the Goddard code.
This was reported to Goddard as a potential bug, and was confirmed to
be a bug in the reference implementation, so it should be turned
on by default.