use shelve, change __exit__ error handling
Use library.shelve instead of __setitem__ and discard

If an error occurs during the with context model usage,
instead of using a ``raise from`` (which will show the borrow
error during pytest and not the original exception) don't
raise a borrow error. This should make test failures easier
to read.